+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## master (unreleased)
 
 * Fix `SpaceAfterPropertyColon` to report nested properties
+* Add `PropertyUnits` allowing customization of which units are allowed for
+  specific properties
 
 ## 0.35.0